Discussed in: http://lists.opencsw.org/pipermail/maintainers/2009-March/001576.html
I've noticed that svn:externals use a short version of SourceForge domain,
sf.net. It causes svn to complain about non-matching SSL certificates and asks
for confirmation when checking out gar svn:externals. I have a prepared svn
change which changes all instances of sf.net into sourceforge.net. I'm
not aware of any major problems that it might cause. I tested it locally
and it Just Works, "svn update" completes successfully after the change.

The change is small in a sense that it boils down to sedding through
svn:externals and replacing sf.net with sourceforge.net. The change is big in a
sense that it touches most of <pkg>/trunk directories.

This change might be confusing if there is a 'trunk' directory with locally
modified svn:externals property, due to, for instance, migration from mGAR v1 to
v2. In such a case, the trunk directory needs to be reverted (non-recursively!)
and svn update needs to be called. Then the property change can be applied
again.


git-svn-id: https://gar.svn.sourceforge.net/svnroot/gar/csw/mgar/pkg@3937 d3b55034-1cff-0310-a425-aefe953e1e90