Initial commit. It builds. More work needed.
It provides the same binaries as mysql does, so we can't build it the same way
as we used to build mysql. Maybe we can rename the binaries, or provide
a suffix. I would hate to build it into a different installation prefix. Or
maybe we can replace MySQL with MariaDB entirely in a future release.


git-svn-id: https://gar.svn.sourceforge.net/svnroot/gar/csw/mgar/pkg@20293 d3b55034-1cff-0310-a425-aefe953e1e90